+++ internal/app/n3dr/artifactsv2/artifactsv2.go
@@ -1,0 +1,9 @@
+package artifactsv2
+
+const (
+	projectName           = "n3dr"
+	CfgFileExt            = "yml"
+	HiddenN3DR            = "." + projectName
+	DefaultCfgFile        = "config"
+	DefaultCfgFileWithExt = DefaultCfgFile + "." + CfgFileExt
+)