+++ docs/CHANGELOG.md
@@ -2,12 +2,36 @@
 ## [Unreleased]
 
 
+<a name="7.0.1"></a>
+## [7.0.1] - 2022-12-18
+### Fix
+- [[#294](https://github.com/030/n3dr/issues/294)] Apply auto updates.
+
+
+<a name="7.0.0"></a>
+## [7.0.0] - 2022-12-17
+### Build
+- **snap:** [[#290](https://github.com/030/n3dr/issues/290)] Stable version 7.
+
+### Fix
+- [[#263](https://github.com/030/n3dr/issues/263)] Remove deprecated commands.
+- **logging:** [[#270](https://github.com/030/n3dr/issues/270)] Optional write to syslog and/or file and default loglevel set to info.
+- **snap:** [[#290](https://github.com/030/n3dr/issues/290)] Snap and release version 7.
+
+### BREAKING CHANGE
+
+The `--debug` and `-d` shorthand have been replaced by `--logLevel debug`.
+
+The `backup`, `upload` and `repositories` commands have been removed.
+
+
 <a name="6.8.3"></a>
-## [6.8.3] - 2022-11-21
+## [6.8.3] - 2022-11-28
 ### Build
 - **auto-updater:** Update schedule.
 
 ### Fix
+- [[#254](https://github.com/030/n3dr/issues/254)] Broken snapcraft build.
 - **maven2:** [[#254](https://github.com/030/n3dr/issues/254)] Upload snapshots.
 
 
@@ -277,7 +301,9 @@
 <a name="1.0.0"></a>
 ## 1.0.0 - 2019-05-12
 
-[Unreleased]: https://github.com/030/n3dr/compare/6.8.3...HEAD
+[Unreleased]: https://github.com/030/n3dr/compare/7.0.1...HEAD
+[7.0.1]: https://github.com/030/n3dr/compare/7.0.0...7.0.1
+[7.0.0]: https://github.com/030/n3dr/compare/6.8.3...7.0.0
 [6.8.3]: https://github.com/030/n3dr/compare/6.8.2...6.8.3
 [6.8.2]: https://github.com/030/n3dr/compare/6.8.1...6.8.2
 [6.8.1]: https://github.com/030/n3dr/compare/6.8.0...6.8.1